requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
django-tinymce==4.0.0
pillow==10.3.0
pytils==0.4.1
sqlparse==0.5.0
